Rename common to constants.
