--- lib/common/app.dart
@@ -1,30 +1,0 @@
-// Flutter imports:
-import 'package:flutter/material.dart';
-
-// All the contants
-const darkGold = Color(0xFFBE830E);
-const brickRed = Color(0xFFD89E7A);
-const lightGreen = Color.fromARGB(255, 120, 219, 137);
-const lightBlue = Color(0xFF61B2CE);
-const exLightBlue = Color(0xFFD8ECF3);
-const darkCopper = Color(0xFFBE4E0E);
-const titleAsh = Color(0xFF30384D);
-const backgroundWhite = Color(0xFFF5F6FC);
-const lightGray = Color(0xFF8793B2);
-const bgOffWhite = Color(0xFFF2F4FC);
-
-const kDefaultPadding = 20.0;
-
-double screenWidth(BuildContext context) => MediaQuery.of(context).size.width;
-double screenHeight(BuildContext context) => MediaQuery.of(context).size.height;
-
-
-// // Folders
-// const List FOLDERS = [
-//   MAIN_RES_DIR,
-//   '$MAIN_RES_DIR/$SHARING_DIR',
-//   '$MAIN_RES_DIR/$SHARED_DIR',
-//   '$MAIN_RES_DIR/$MED_DIR',
-//   '$MAIN_RES_DIR/$ENC_DIR',
-//   '$MAIN_RES_DIR/$LOGS_DIR'
-// ];
