+++ lib/constants/rest_api.dart
@@ -1,13 +1,34 @@
-// Flutter imports.
-import 'dart:async';
+/// DESCRIPTION
+///
+/// Copyright (C) 2023, Software Innovation Institute
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// Time-stamp: <Wednesday 2023-11-01 08:26:39 +1100 Graham Williams>
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://www.gnu.org/licenses/>.
+///
+/// Authors: AUTHORS
 
-// Package imports.
-import 'package:solid_auth/solid_auth.dart';
-import 'package:jwt_decoder/jwt_decoder.dart';
+// import 'dart:async';
 
-// Project imports.
-import 'package:podnotes/common/app.dart';
+// import 'package:solid_auth/solid_auth.dart';
+// import 'package:jwt_decoder/jwt_decoder.dart';
 
+// import 'package:podnotes/common/app.dart';
 
 // Future<List> initialStructureTest(Map authData) async {
 //   var rsaInfo = authData['rsaInfo'];
